ssl_requirement.txt
zstd==1.4.9=haebb681_0
zope.interface==5.4.0=py39h7f8727e_0
zope.event==4.5.0=py39h06a4308_0
zope==1.0=py39h06a4308_1
zlib==1.2.11=h7b6447c_3
zfp==0.5.5=h2531618_6
zeromq==4.3.4=h2531618_0
yaml==0.2.5=h7b6447c_0
xz==5.2.5=h7b6447c_0
xlwt==1.3.0=py39h06a4308_0
wurlitzer==2.1.1=py39h06a4308_0
wrapt==1.12.1=py39he8ac12f_1
widgetsnbextension==3.5.1=py39h06a4308_0
webencodings==0.5.1=py39h06a4308_1
watchdog==2.1.3=py39h06a4308_0
unixodbc==2.3.9=h7b6447c_0
unicodecsv==0.14.1=py39h06a4308_0
ujson==4.0.2=py39h2531618_0
typed-ast==1.4.3=py39h7f8727e_1
tornado==6.1=py39h27cfd23_0
torchvision==0.11.3=py39_cu113
torchaudio==0.10.2=py39_cu113
tk==8.6.11=h1ccaba5_0
terminado==0.9.4=py39h06a4308_0
tbb4py==2021.4.0=py39hd09550d_0
tbb==2021.4.0=hd09550d_0
sympy==1.9=py39h06a4308_0
statsmodels==0.12.2=py39h27cfd23_0
sqlite==3.36.0=hc218d9a_0
sqlalchemy==1.4.22=py39h7f8727e_0
spyder-kernels==2.1.3=py39h06a4308_0
spyder==5.1.5=py39h06a4308_1
sphinxcontrib==1.0=py39h06a4308_1
sniffio==1.2.0=py39h06a4308_1
snappy==1.1.8=he6710b0_0
sip==4.19.13=py39h2531618_0
simplegeneric==0.8.1=py39h06a4308_2
setuptools==58.0.4=py39h06a4308_0
secretstorage==3.3.1=py39h06a4308_0
scipy==1.7.1=py39h292c36d_2
scikit-learn-intelex==2021.3.0=py39h06a4308_0
scikit-learn==0.24.2=py39ha9443f7_0
scikit-image==0.18.3=py39h51133e4_0
ruamel_yaml==0.15.100=py39h27cfd23_0
rtree==0.9.7=py39h06a4308_1
ripgrep==12.1.1=0
regex==2021.8.3=py39h7f8727e_0
readline==8.1=h27cfd23_0
qt==5.9.7=h5867ecd_1
pyzmq==22.2.1=py39h295c915_1
pyyaml==6.0=py39h7f8727e_1
pywavelets==1.1.1=py39h6323ea4_4
pytorch==1.10.2=py3.9_cuda11.3_cudnn8.2.0_0
python==3.9.7=h12debd9_1
pytest==6.2.4=py39h06a4308_2
pytables==3.6.1=py39h77479fe_1
pysocks==1.7.1=py39h06a4308_0
pyrsistent==0.18.0=py39heee7806_0
pyqt==5.9.2=py39h2531618_6
pyodbc==4.0.31=py39h295c915_0
pylint==2.9.6=py39h06a4308_1
pyerfa==2.0.0=py39h27cfd23_0
pycurl==7.44.1=py39h8f2d780_1
pycosat==0.6.3=py39h27cfd23_0
py-lief==0.10.1=py39h2531618_1
psutil==5.8.0=py39h27cfd23_1
protobuf==3.15.8=py39he80948d_0
ply==3.11=py39h06a4308_0
pluggy==0.13.1=py39h06a4308_0
pkginfo==1.7.1=py39h06a4308_0
pixman==0.40.0=h7f8727e_1
pip==21.2.4=py39h06a4308_0
pillow==8.4.0=py39h5aabda8_0
pep8==1.7.1=py39h06a4308_0
pcre==8.45=h295c915_0
patsy==0.5.2=py39h06a4308_0
pathlib2==2.3.6=py39h06a4308_2
path==16.0.0=py39h06a4308_0
patchelf==0.13=h295c915_0
pango==1.45.3=hd140c19_0
pandocfilters==1.4.3=py39h06a4308_1
pandas==1.3.4=py39h8c16a72_0
openssl==1.1.1l=h7f8727e_0
openjpeg==2.4.0=h3ad879b_0
openh264==2.1.1=h780b84a_0
numpy-base==1.20.3=py39h74d4b33_0
numpy==1.20.3=py39hf144106_0
numexpr==2.7.3=py39h22e1b3c_1
numba==0.54.1=py39h51133e4_0
notebook==6.4.5=py39h06a4308_0
nettle==3.7.3=hbbd107a_1
ncurses==6.3=heee7806_1
nbconvert==6.1.0=py39h06a4308_0
navigator-updater==0.2.1=py39_1
mypy_extensions==0.4.3=py39h06a4308_0
multipledispatch==0.6.0=py39h06a4308_0
msgpack-python==1.0.2=py39hff7bd54_1
mpmath==1.2.1=py39h06a4308_0
mpich==3.3.2=hc856adb_0
mpfr==4.0.2=hb69a4c5_1
mpc==1.1.0=h10f8cd9_1
mkl_random==1.2.2=py39h51133e4_0
mkl_fft==1.3.1=py39hd3c417c_0
mkl-service==2.4.0=py39h7f8727e_0
mkl==2021.4.0=h06a4308_640
mistune==0.8.4=py39h27cfd23_1000
mccabe==0.6.1=py39h06a4308_1
matplotlib-base==3.4.3=py39hbbc1b5f_0
matplotlib==3.4.3=py39h06a4308_0
markupsafe==1.1.1=py39h27cfd23_0
lzo==2.10=h7b6447c_2
lz4-c==1.9.3=h295c915_1
lxml==4.6.3=py39h9120a33_0
locket==0.2.1=py39h06a4308_1
llvmlite==0.37.0=py39h295c915_1
libzopfli==1.0.3=he6710b0_0
libxslt==1.1.34=hc22bd24_0
libxml2==2.9.12=h03d6c58_0
libxcb==1.14=h7b6447c_0
libwebp-base==1.2.0=h27cfd23_0
libwebp==1.2.0=h89dd481_0
libuv==1.40.0=h7b6447c_0
libuuid==1.0.3=h7f8727e_2
libunistring==0.9.10=h7f98852_0
libtool==2.4.6=h7b6447c_1005
libtiff==4.2.0=h85742a9_0
libtasn1==4.16.0=h27cfd23_0
libstdcxx-ng==9.3.0=hd4cf53a_17
libssh2==1.9.0=h1ba5d50_1
libspatialindex==1.9.3=h2531618_0
libsodium==1.0.18=h7b6447c_0
libprotobuf==3.15.8=h780b84a_0
libpng==1.6.37=hbc83047_0
libnghttp2==1.41.0=hf8bcb03_2
libllvm11==11.1.0=h3826bc1_0
liblief==0.10.1=h2531618_1
libidn2==2.3.2=h7f98852_0
libiconv==1.15=h516909a_1006
libgomp==9.3.0=h5101ec6_17
libgfortran4==7.5.0=ha8ba4b0_17
libgfortran-ng==7.5.0=ha8ba4b0_17
libgcc-ng==9.3.0=h5101ec6_17
libffi==3.3=he6710b0_2
libev==4.33=h7f8727e_1
libedit==3.1.20210910=h7f8727e_0
libdeflate==1.8=h7f8727e_5
libcurl==7.78.0=h0b77cf5_0
libarchive==3.4.2=h62408e4_0
libaec==1.0.4=he6710b0_1
lerc==3.0=h295c915_0
ld_impl_linux-64==2.35.1=h7274673_9
lcms2==2.12=h3be6417_0
lazy-object-proxy==1.6.0=py39h27cfd23_0
lame==3.100=h7f98852_1001
krb5==1.19.2=hac12032_0
kiwisolver==1.3.1=py39h2531618_0
keyring==23.1.0=py39h06a4308_0
jxrlib==1.1=h7b6447c_2
jupyter_server==1.4.1=py39h06a4308_0
jupyter_core==4.8.1=py39h06a4308_0
jupyter==1.0.0=py39h06a4308_7
jpeg==9d=h7f8727e_0
jedi==0.18.0=py39h06a4308_1
jbig==2.1=hdba287a_0
ipython==7.29.0=py39hb070fc8_0
ipykernel==6.4.1=py39h06a4308_1
intel-openmp==2021.4.0=h06a4308_3561
inflection==0.5.1=py39h06a4308_0
importlib-metadata==4.8.1=py39h06a4308_0
imagecodecs==2021.8.26=py39h4cda21f_0
icu==58.2=he6710b0_3
hdf5==1.10.6=hb1b8bf9_0
harfbuzz==2.8.1=h6f93f22_0
h5py==3.3.0=py39h930cdd6_0
gstreamer==1.14.0=h28cd5cc_2
gst-plugins-base==1.14.0=h8213a91_2
greenlet==1.1.1=py39h295c915_0
graphite2==1.3.14=h23475e2_0
gnutls==3.6.15=he1e5248_0
gmpy2==2.0.8=py39h8083e48_3
gmp==6.2.1=h2531618_2
glib==2.69.1=h5202010_0
giflib==5.2.1=h7b6447c_0
gevent==21.8.0=py39h7f8727e_1
gettext==0.19.8.1=h0b5b191_1005
get_terminal_size==1.0.0=haa9412d_0
fribidi==1.0.10=h7b6447c_0
freetype==2.10.4=h5ab3b9f_0
fontconfig==2.13.1=h6c09931_0
ffmpeg==4.3=hf484d3e_0
fastcache==1.1.0=py39he8ac12f_0
expat==2.4.1=h2531618_2
et_xmlfile==1.1.0=py39h06a4308_0
entrypoints==0.3=py39h06a4308_0
docutils==0.17.1=py39h06a4308_1
distributed==2021.10.0=py39h06a4308_0
debugpy==1.4.1=py39h295c915_0
dbus==1.13.18=hb2f20db_0
dal==2021.3.0=h06a4308_557
daal4py==2021.3.0=py39hae6d005_0
cytoolz==0.11.0=py39h27cfd23_0
cython==0.29.24=py39hdbfa776_0
cycler==0.10.0=py39h06a4308_0
curl==7.78.0=h1ccaba5_0
cudatoolkit==11.3.1=h2bc3f7f_2
cryptography==3.4.8=py39hd23ed53_0
conda-package-handling==1.7.3=py39h3811e60_0
conda-build==3.21.5=py39h06a4308_0
conda==4.11.0=py39hf3d152e_2
clyent==1.2.2=py39h06a4308_1
charls==2.2.0=h2531618_0
chardet==4.0.0=py39h06a4308_1003
cfitsio==3.470=hf0d0db6_6
cffi==1.14.6=py39h400218f_0
certifi==2021.10.8=py39h06a4308_0
cairo==1.16.0=hf32fb01_1
ca-certificates==2021.10.26=h06a4308_2
c-ares==1.17.1=h27cfd23_0
bzip2==1.0.8=h7b6447c_0
brunsli==0.1=h2531618_0
brotlipy==0.7.0=py39h27cfd23_1003
brotli==1.0.9=he6710b0_2
bottleneck==1.3.2=py39hdd57654_1
boto==2.49.0=py39h06a4308_0
bokeh==2.4.1=py39h06a4308_0
blosc==1.21.0=h8c45485_0
blas==1.0=mkl
bkcharts==0.2=py39h06a4308_0
bitarray==2.3.0=py39h7f8727e_1
astropy==4.3.1=py39h09021b7_0
astroid==2.6.6=py39h06a4308_0
arrow==0.13.1=py39h06a4308_0
argon2-cffi==20.1.0=py39h27cfd23_1
argh==0.26.2=py39h06a4308_0
anyio==2.2.0=py39h06a4308_1
anaconda-navigator==2.1.1=py39_0
anaconda-client==1.9.0=py39h06a4308_0
anaconda==2021.11=py39_0
_openmp_mutex==4.5=1_gnu
_libgcc_mutex==0.1=main
_ipyw_jlab_nb_ext_conf==0.1.0=py39h06a4308_1